crud controller validations

+++ src/api/api-principal/mixins/crud-controller.ts
@@ -1,10 +1,26 @@
-import {FindFullQuery, PrincipalService} from '../../..';
+import {BaseDTO, DtoConfig, DtoConfigInterface, FindFullQuery, PrincipalService} from '../../..';
 import {DeepPartial} from 'typeorm';
-import {Body, Delete, Get, Param, Post, Put, Query} from '@nestjs/common';
+import {
+    BadRequestException,
+    Body,
+    Delete,
+    Get,
+    InternalServerErrorException, NotFoundException,
+    Param,
+    Post,
+    Put,
+    Query
+} from '@nestjs/common';
 import {AbstractController} from './abstract-controller';
+import {validateMany} from '../../shared-utils/validate-many';
+import {plainToClass} from 'class-transformer';
+import {validate} from 'class-validator';
 
 
-export function CrudController<T>(): typeof AbstractController {
+export function CrudController<T>(dtoConfig: DtoConfigInterface | DtoConfig = {
+    createDtoType: BaseDTO,
+    updateDtoType: BaseDTO
+}): typeof AbstractController {
     class BaseController extends AbstractController<T> {
 
         constructor(
@@ -14,28 +30,84 @@ export function CrudController<T>(): typeof AbstractController {
         }
 
         @Post()
-        createMany(
-            @Body('records') newRecords: DeepPartial<T>[]): any {
-            return this._service.createMany(newRecords);
+        async createMany(
+            @Body('records') newRecords: DeepPartial<T>[]): Promise<T[]> {
+            const validationErrors = await validateMany(newRecords, dtoConfig.createDtoType);
+            if (validationErrors.length > 0) {
+                console.error(validationErrors);
+                throw new BadRequestException('Bad Request');
+            } else {
+                try {
+                    return this._service.createMany(newRecords);
+                } catch (error) {
+                    console.error(
+                        {
+                            error,
+                            message: 'Error on create',
+                            data: {records: newRecords},
+                        }
+                    );
+                    throw new InternalServerErrorException('Server Error');
+                }
+            }
         }
 
         @Post()
-        createOne(
-            @Body() newRecord: DeepPartial<T>): any {
-            return this._service.createOne(newRecord);
+        async createOne(
+            @Body() newRecord: DeepPartial<T>): Promise<T> {
+            const entityDto = plainToClass(dtoConfig.createDtoType, newRecord) as object;
+            const validationErrors = await validate(entityDto);
+            if (validationErrors.length > 0) {
+                console.error(validationErrors);
+                throw new BadRequestException('Bad Request');
+            } else {
+                try {
+                    return this._service.createOne(newRecord);
+                } catch (error) {
+                    console.error(
+                        {
+                            error,
+                            message: 'Error on create',
+                            data: {record: newRecord},
+                        }
+                    );
+                    throw new InternalServerErrorException('Server Error');
+                }
+            }
         }
 
         @Delete(':id')
-        deleteOne(
+        async deleteOne(
             @Param('id') id: number,
-        ): any {
-            return this._service.deleteOne(id);
+        ): Promise<T> {
+            const isIdValid = this.validateId(id);
+            if (isIdValid) {
+                try {
+                    await this._service.findOneById(id);
+                } catch (error) {
+                    throw new NotFoundException({message: 'Record not found'});
+                }
+                try {
+                    return await this._service.deleteOne(id);
+                } catch (error) {
+                    console.error(
+                        {
+                            error,
+                            message: 'Error on delete',
+                            data: {id},
+                        },
+                    );
+                    throw new InternalServerErrorException({message: 'Server Error'});
+                }
+            } else {
+                throw new BadRequestException({message: 'Invalid Id'});
+            }
         }
 
         @Get()
         async findAll(
             @Query('query') searchCriteria: any,
-        ) {
+        ): Promise<{ nextQuery: any; data: T[]; total: number }> {
             let result: [T[], number];
             let skip = 0;
             let take = 10;
@@ -50,26 +122,28 @@ export function CrudController<T>(): typeof AbstractController {
                     query = {where: {}, skip: 0, take: 10};
                     result = await this._service.findAll({} as FindFullQuery);
                 }
-                const total = +result[1];
-                const rest = total - (skip + take);
-                const isLastPage = rest <= 0;
-                let nextQuery = null;
+                const totalRecords: number = +result[1];
+                const data: T[] = result[0];
+                const restingRecords: number = totalRecords - (skip + take);
+                const isLastPage: boolean = restingRecords <= 0;
+                let nextQuery: any = null;
                 if (!isLastPage) {
-                    const isNotLimit = rest >= take;
+                    const isNotLimit = restingRecords >= take;
                     const nextSkip = skip + take;
-                    const nextTake = isNotLimit ? take : rest;
+                    const nextTake = isNotLimit ? take : restingRecords;
                     const partialQuery: Partial<FindFullQuery> = {...query};
                     partialQuery.skip = nextSkip;
                     partialQuery.take = nextTake;
                     if (query.where) {
-                        partialQuery.where = Object.keys(query.where).length > 0 ? partialQuery.where : undefined;
+                        const hasWhereConditions = Object.keys(query.where).length;
+                        partialQuery.where = hasWhereConditions > 0 ? partialQuery.where : undefined;
                     }
                     nextQuery = partialQuery;
                 }
                 return {
                     nextQuery,
-                    data: result[0],
-                    total: result[1],
+                    data,
+                    total: totalRecords,
                 };
             } catch (error) {
                 console.error(
@@ -89,18 +163,67 @@ export function CrudController<T>(): typeof AbstractController {
         }
 
         @Get(':id')
-        findOneById(
+        async findOneById(
             @Param('id') id: number,
-        ): any {
-            return this._service.findOneById(id);
+        ): Promise<T> {
+            const isIdValid = this.validateId(id);
+            if (isIdValid) {
+                try {
+                    return this._service.findOneById(Number(id),);
+                } catch (error) {
+                    console.error(
+                        {
+                            error,
+                            mensaje: 'Error on fetch results',
+                            data: {id},
+                        },
+                    );
+                    throw new InternalServerErrorException({message: 'Server Error'});
+                }
+            } else {
+                throw new BadRequestException({message: 'Invalid Id'});
+            }
         }
 
         @Put(':id')
-        updateOne(
+        async updateOne(
             @Body() recordToUpdate: DeepPartial<T>,
-            @Param('id') id: number
-        ): any {
-            return this._service.updateOne(id, recordToUpdate);
+            @Param('id') id: number,
+        ): Promise<T> {
+            const isValidId = this.validateId(id);
+            if (isValidId) {
+                const dtoEntity = plainToClass(dtoConfig.updateDtoType, recordToUpdate) as object;
+                const validationErrors = await validate(dtoEntity);
+                if (validationErrors.length > 0) {
+                    console.error(validationErrors);
+                    throw new BadRequestException({message: 'Bad Request'});
+                } else {
+                    try {
+                        await this._service.findOneById(id);
+                    } catch (error) {
+                        throw new NotFoundException({message: 'Record not found'});
+                    }
+                    try {
+                        return this._service.updateOne(id, recordToUpdate);
+                    } catch (error) {
+                        console.error(
+                            {
+                                error,
+                                message: 'Error al actualizar',
+                                data: {id, datosActualizar: recordToUpdate},
+                            }
+                        );
+                        throw new InternalServerErrorException({message: 'Server Error'});
+                    }
+
+                }
+            } else {
+                throw new BadRequestException({message: 'Invalid Id'});
+            }
+        }
+
+        protected validateId(id: any): boolean {
+            return !isNaN(Number(id));
         }
     }
